+++ src/services/audio.service.ts
@@ -1,10 +1,15 @@
 const player = require("play-sound")();
 import * as path from "path";
+import * as vscode from "vscode";
 
 export class AudioService {
   private audioPath = path.join(__dirname, "..", "..", "medias", "audios");
 
   playSuccess() {
+    if (!this.isSoundEnabled()) {
+      return;
+    }
+
     try {
       player.play(path.join(this.audioPath, "success.mp3"));
     } catch (e) {
@@ -13,10 +18,18 @@ export class AudioService {
   }
 
   playError() {
+    if (!this.isSoundEnabled()) {
+      return;
+    }
+    
     try {
       player.play(path.join(this.audioPath, "error.mp3"));
     } catch (e) {
       console.log(`impossible to play sound: ${e}`);
     }
+  }
+
+  private isSoundEnabled(): boolean {
+    return vscode.workspace.getConfiguration("koby").get('enableSound', true);
   }
 }